Maybe final step

+++ ewm-service/src/main/java/ru/practicum/ewm/categories/CategoryServiceImpl.java
@@ -1,7 +1,6 @@
 package ru.practicum.ewm.categories;
 
 import lombok.RequiredArgsConstructor;
-import org.hibernate.engine.jdbc.spi.SqlExceptionHelper;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import ru.practicum.ewm.categories.dto.CategoryDto;
@@ -9,9 +8,10 @@ import ru.practicum.ewm.categories.dto.NewCategoryDto;
 import ru.practicum.ewm.categories.interfaces.CategoryRepository;
 import ru.practicum.ewm.categories.interfaces.CategoryService;
 import ru.practicum.ewm.categories.model.Category;
+import ru.practicum.ewm.events.interfaces.EventRepository;
+import ru.practicum.ewm.exception.ConflictRequestException;
 import ru.practicum.ewm.exception.DataNotFoundRequestException;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -20,6 +20,7 @@ import java.util.stream.Collectors;
 public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     @Override
     public CategoryDto addCategory(NewCategoryDto newCategoryDto) {
@@ -32,6 +33,10 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public void deleteCategory(Long catId) {
 
+        if (eventRepository.existsByCategoryId(catId)) {
+            throw new ConflictRequestException("Cannot delete Category with id = " + catId +
+                    " because exists events with this Category");
+        }
         categoryRepository.deleteById(catId);
     }
